### 2021-09-03 #3690 @whedon
[PRE REVIEW]: gdess: A framework for evaluating simulated atmospheric CO~2~ in Earth System Models
**Submitting author:** @dkauf42 (<a href="http://orcid.org/0000-0002-1487-7298">Daniel E Kaufman</a>)
**Repository:** https://github.com/E3SM-Project/gdess
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.beta1
