### 2021-09-03 #3690 @kyleniemeyer
Hi @dkauf42 thanks for your submission to JOSS! Unfortunately, we don't have an available editor in this area right now, so I have to put this on our waitlist. As soon as someone becomes available, we will assign it and get moving.

In the meantime, could you review and resolve those DOI issues in the article? And, any reviewer suggestions you can give would be welcome.
