### 2021-09-05 a99efc8
standardize references: doi formats, remove uncited, make style more consistent
