+++ paper/paper.md
@@ -39,19 +39,19 @@ We introduce `gdess` (a.k.a., Greenhouse gas Diagnostics for Earth System Simula
 
 # Statement of need
 
-Thorough evaluation of simulated atmospheric CO~2~ concentrations—by comparing against observations—requires multiple diagnostics, metrics, and visualizations. During the past decade, such evaluations have utilized certain common methods, such as aggregating in situ measurements into latitude bands and detrending of multidecadal time series to investigate seasonal cycles [@Keppel-Aleks2013b; @Ott2015b; @Liptak2017b; @Jing2018b; @Chevallier2019e; @Weir2020b]. However, the construction of diagnostics used in these evaluations has not been automated in an open-source tool available to the broader atmospheric modeling community. Thus, each modeling or analysis team has had to decide on and code their own preferred set of diagnostics, resulting in redundancies and potential inconsistencies among efforts.
+Thorough evaluation of simulated atmospheric CO~2~ concentrations—by comparing against observations—requires multiple diagnostics, metrics, and visualizations. During the past decade, such evaluations have utilized certain common methods, such as aggregating in situ measurements into latitude bands and detrending of multidecadal time series to investigate seasonal cycles [@KeppelAleks2013b; @Ott2015b; @Liptak2017b; @Jing2018b; @Chevallier2019e; @WeirEtAl2021acp]. However, the construction of diagnostics used in these evaluations has not been automated in an open-source tool available to the broader atmospheric modeling community. Thus, each modeling or analysis team has had to decide on and code their own preferred set of diagnostics, resulting in redundancies and potential inconsistencies among efforts.
 
-Several software packages have been developed to streamline the application of diagnostics for ESM benchmarking. These tools share related functionality with `gdess`, and some have directly inspired the `gdess` design and our development approach. For example, the ESM Evaluation Tool (ESMValTool; Eyring et al., 2016; 2020) has been used to generate specific figures from the literature—including a comparison of column-averaged CO~2~ values as performed by Gier et al. (2020)—but does not provide for tailored processing of varied CO~2~ data sources. We adopted the term *recipe* from its use by ESMValTool. The International Land Model Benchmarking (ILAMB) System [@Collier2018a] excels at intercomparisons between multiple land models and has been used to benchmark inferred CO~2~ concentrations against surface station measurements [@WuEtAl2020jhydro]. In contrast to `gdess`, ILAMB provides the means to evaluate emulated results but not prognostic simulations for CO~2~ [@keppelaleks2021PersComm].
+Several software packages have been developed to streamline the application of diagnostics for ESM benchmarking. These tools share related functionality with `gdess`, and some have directly inspired the `gdess` design and our development approach. For example, the ESM Evaluation Tool (ESMValTool; @Eyring2016c; @Eyring2020b) has been used to generate specific figures from the literature—including a comparison of column-averaged CO~2~ values as performed by Gier et al. (2020)—but does not provide for tailored processing of varied CO~2~ data sources. We adopted the term *recipe* from its use by ESMValTool. The International Land Model Benchmarking (ILAMB) System [@Collier2018a] excels at intercomparisons between multiple land models and has been used to benchmark inferred CO~2~ concentrations against surface station measurements [@WuEtAl2020jhydro]. In contrast to `gdess`, ILAMB provides the means to evaluate emulated results but not prognostic simulations for CO~2~ [@KeppelAleks2021PersComm].
 
 # Design and data sources
 
[…]
