+++ paper/paper.bib
@@ -1,156 +1,100 @@
 ############################################################
 ## BOOKS
-
-@BOOK{VanRossumAndDrake2009,
- author = {Van Rossum, Guido and Drake, Fred L.},
- title = {Python 3 Reference Manual},
- year = {2009},
- isbn = {1441412697},
- publisher = {CreateSpace},
- address = {Scotts Valley, CA}
+@book{VanRossumAndDrake2009,
+	title        = {{Python 3 Reference Manual}},
+	author       = {Van Rossum, Guido and Drake, Fred L.},
+	year         = 2009,
+	publisher    = {CreateSpace},
+	address      = {Scotts Valley, CA},
+	isbn         = 1441412697,
 }
-
-##########################################################
-## MISC
-
-@misc{keppelaleks2021PersComm,
-  author = {Keppel-Aleks, Gretchen},
-  year={2021},
-  date = "2021-04-28",
-  howpublished = {personal communication}
-}
-
-@misc{KeelingEtAl2001sioref,
-author = {Keeling, C.D. and Piper, S.C. and Bacastow, R.B. and Wahlen, M. and Whorf, T.P. and Heimann, M. and Meijer, H.A.},
-year = {2001},
-pages = {88},
-date = "2001-06-01",
-title = {Exchanges of atmospheric CO2 and 13CO2 with the terrestrial biosphere and oceans from 1978 to 2000.  I. Global aspects},
-url = {https://escholarship.org/uc/item/09v319r9},
-editor = {Scripps  Institution of Oceanography, San Diego},
-publisher = {UC San Diego: Library – Scripps Digital},
-series = {SIO Reference Series}, 
-number = {01-06}, 
-}
-
-##########################################################
-## ARTICLES
-
-@article{Fer2021gcb,
-author = {Fer, Istem and Gardella, Anthony K. and Shiklomanov, Alexey N. and Campbell, Eleanor E. and Cowdery, Elizabeth M. and De Kauwe, Martin G. and Desai, Ankur and Duveneck, Matthew J. and Fisher, Joshua B. and Haynes, Katherine D. and Hoffman, Forrest M. and Johnston, Miriam R. and Kooper, Rob and LeBauer, David S. and Mantooth, Joshua and Parton, William J. and Poulter, Benjamin and Quaife, Tristan and Raiho, Ann and Schaefer, Kevin and Serbin, Shawn P. and Simkins, James and Wilcox, Kevin R. and Viskari, Toni and Dietze, Michael C.},
-title = {Beyond ecosystem modeling: A roadmap to community cyberinfrastructure for ecological data-model integration},
-journal = {Global Change Biology},
-volume = {27},
-number = {1},
-pages = {13-26},
-keywords = {accessibility, benchmarking, community cyberinfrastructure, data, data assimilation, ecosystem models, interoperability, reproducibility},
-doi = {https://doi.org/10.1111/gcb.15409},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/gcb.15409},
-eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/gcb.15409},
[…]
