+++ .github/workflows/lint_and_test.yml
@@ -4,12 +4,12 @@
 name: Tests
 
 on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
+  #  push:
+  #    branches:
+  #      - main
+  #  pull_request:
+  #    branches:
+  #      - main
   workflow_dispatch:
     inputs:
       debug_enabled:
