### 2021-09-08 #3690 @dkauf42
Hi @kyleniemeyer, thank you for the information. I just updated the paper files to resolve DOI issues.

For reviewers, I can suggest:

- Dr. Bianca Baier (U.S. National Oceanic and Atmos. Admin.; email: Bianca.Baier@noaa.gov)
- Dr. Lesley Ott (U.S. Nat'l Aeronautics and Space Admin.; email: lesley.e.ott@nasa.gov)
- Dr. Samantha Basile (U.S. National Climate Assessment; webpage: https://www.globalchange.gov/staff/samantha-basile)
- Dr. Havala Pye (U.S. Env. Protection Agency; email: pye.havala@epa.gov; GitHub: havalapye)
- Dr. Jesse Bash (U.S. Env. Protection Agency; email: bash.jesse@epa.gov; GitHub: jessebash)
- Dr. Elizabeth Adams (U.S. Env. Protection Agency; GitHub: lizadams; email: Adams.elizabeth@epa.gov)
- Doctoral candidate, Karolina Pantazatou (Lund University; email: karolina.pantazatou@nateko.lu.se; GitHub: KarolinaPntzt)
