### 2021-09-23 #3690 @kthyng
@richardjgowers is this a reasonable fit for you?
