### 2021-10-03 #3690 @richardjgowers
@KarolinaPntzt @lizadams @havalapye Hi 👋 would any of you be available to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
