### 2021-12-03 #3690 @dkauf42
Hi @richardjgowers, just checking back in since the last visible comment is from two months ago: Are there any updates on our manuscript's status?  Thanks!
