### 2022-01-05 #3690 @BunnyVon
@richardjgowers Happy New Year! I'm one of the co-authors of this manuscript. I wonder if you have any updates on the review process.
There doesn't appear to have been any update for 3 months. So we also wonder if there is any additional information we can provide to you — to help advance our manuscript to the reviewing stage. Thanks :)
