### 2022-02-16 #3690 @dhhagan
@dkauf42  First, let me apologize for the long delay at the pre-review stage. After an initial look/review, I made the decision to ask other JOSS editors to evaluate whether or not this submission meets the JOSS guideline for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort). I will let you know as soon as that is resolved. Again, sorry for the long delay!
