### 2022-04-07 #3690 @rabernat
I do not have room in my schedule to review right now. However, I will tag a few people I know who are fluent in both Python and CO2 cycles: @sridge, @lgloege. Maybe one of them can step in.
