### 2022-04-20 61819eb
add set_path_vars.bat for Windows OS and describe its use in the readme
