### 2022-05-02 #4326 @slayoo
@dkauf42, re the following requirement:
> Does the full list of paper authors seem appropriate and complete?

According to https://github.com/E3SM-Project/gdess/graphs/contributors, you are the only code contributor. I suggest thus adding to the paper an "Author contributions" section commenting on each authors' contributions. See also https://joss.readthedocs.io/en/latest/submitting.html#authorship
