### 2022-05-03 2a8557c
add author contribution section to paper.md
