### 2022-05-03 #4326 @slayoo
@dkauf42 Isn't there a naming mismatch here? JOSS paper and the GitHub repo is named `gdess`. The `README.md` mentions `pip uninstall co2_diag` but the `setup.cfg` file lists `name = e3sm_co2_diag`. All three names seem to be available on pypi.org. Why not to stick to one & register it and upload to PyPI (and adjust `setup.cfg` and `README.md` accordingly?
