+++ gdess/data_source/models/e3sm/collection.py
@@ -1,12 +1,12 @@
-from co2_diag import set_verbose, benchmark_recipe
-from co2_diag.data_source.models.e3sm.calculation import getPMID
-from co2_diag.data_source.multiset import Multiset
-from co2_diag.operations.datasetdict import DatasetDict
-from co2_diag.operations.time import to_datetimeindex
-from co2_diag.operations.convert import co2_kgfrac_to_ppm
-from co2_diag.graphics.utils import aesthetic_grid_no_spines, mysavefig
-from co2_diag.formatters import append_before_extension
-from co2_diag.recipe_parsers import add_shared_arguments_for_recipes, parse_recipe_options
+from gdess import set_verbose, benchmark_recipe
+from gdess.data_source.models.e3sm.calculation import getPMID
+from gdess.data_source.multiset import Multiset
+from gdess.operations.datasetdict import DatasetDict
+from gdess.operations.time import to_datetimeindex
+from gdess.operations.convert import co2_kgfrac_to_ppm
+from gdess.graphics.utils import aesthetic_grid_no_spines, mysavefig
+from gdess.formatters import append_before_extension
+from gdess.recipe_parsers import add_shared_arguments_for_recipes, parse_recipe_options
 import xarray as xr
 import matplotlib.pyplot as plt
 from typing import Union
