+++ gdess/data_source/observations/gvplus_surface.py
@@ -1,13 +1,13 @@
-from co2_diag import set_verbose, load_stations_dict, load_config_file, benchmark_recipe
-from co2_diag.data_source.observations.load import load_data_with_regex, dataset_from_filelist
-from co2_diag.data_source.multiset import Multiset
-from co2_diag.operations.datasetdict import DatasetDict
-from co2_diag.operations.time import select_between, ensure_dataset_datetime64, ensure_datetime64_array
-from co2_diag.operations.convert import co2_molfrac_to_ppm
-from co2_diag.graphics.single_source_plots import plot_annual_series
-from co2_diag.graphics.utils import aesthetic_grid_no_spines, mysavefig
-from co2_diag.recipe_parsers import add_shared_arguments_for_recipes, parse_recipe_options
-from co2_diag.formatters import append_before_extension
+from gdess import set_verbose, load_stations_dict, load_config_file, benchmark_recipe
+from gdess.data_source.observations.load import load_data_with_regex, dataset_from_filelist
+from gdess.data_source.multiset import Multiset
+from gdess.operations.datasetdict import DatasetDict
+from gdess.operations.time import select_between, ensure_dataset_datetime64, ensure_datetime64_array
+from gdess.operations.convert import co2_molfrac_to_ppm
+from gdess.graphics.single_source_plots import plot_annual_series
+from gdess.graphics.utils import aesthetic_grid_no_spines, mysavefig
+from gdess.recipe_parsers import add_shared_arguments_for_recipes, parse_recipe_options
+from gdess.formatters import append_before_extension
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
