+++ tests/test_cmip_collection.py
@@ -1,8 +1,8 @@
 import pytest
 
-from co2_diag.data_source.models.cmip.cmip_collection import Collection
-from co2_diag.data_source.models.cmip.cmip_name_utils import matched_model_and_experiment
-from co2_diag.operations.datasetdict import DatasetDict
+from gdess.data_source.models.cmip.cmip_collection import Collection
+from gdess.data_source.models.cmip.cmip_name_utils import matched_model_and_experiment
+from gdess.operations.datasetdict import DatasetDict
 
 
 @pytest.fixture
