### 2022-05-04 1979518
make changes in readme to match revised package name
