### 2022-05-04 a4d06c0
change default environment.yml env name to match revised package name
