### 2022-05-04 c402a3d
change references to cli file in readme to match revised package and file names
