### 2022-05-04 aa5556f
remove deprecated legend handler _legmarker usage
