### 2022-05-04 1232758
rearrange readme to correspond to a pypi installation
