### 2022-05-04 d233c43
add repository urls to setup.cfg
