+++ setup.cfg
@@ -1,11 +1,14 @@
 [metadata]
 name = gdess
+version = 1.0.beta1
 description = Comparison of simulated atmospheric CO2 against observational data
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = Daniel E Kaufman
 author_email = daniel.kaufman@pnnl.gov
-version = 1.0.beta1
+url = https://github.com/E3SM-Project/gdess
+project_urls =
+    Bug Tracker = https://github.com/E3SM-Project/gdess/issues
 classifiers =
   Intended Audience :: Science/Research
   Programming Language :: Python :: 3.8
