### 2022-05-04 f941e8c
update readme to reflect package retrieval from pypi
