### 2022-05-13 #4326 @simonom
@dkauf42, regarding the substantial scholarly effort section of the review.

The ESMValTool is rightly referenced in the statement of need in the paper.  And its overarching objective appears to be the same as gdess, which means the benefit of gdess (over ESMValTool) lies in the details (but tell me if I'm wrong on this).

The distinction between the two software packages is stated in the gdess paper (Statement of need section) as: '—but does not provide for tailored processing of varied CO2 data sources'.  From this statement I cannot tell the benefit of gdess over ESMValTool.

Although I appreciate the need for concision in the JOSS paper, I recommend that the advantage of gdess over ESMValTool be expanded in the Statement of need, so that it is clearer.
