### 2022-05-17 5e5fe47
update output items for surface_trends recipe
- For standard output: 'xdata_obs' and 'xdata_mdl' were added to the output dictionary; and the 'model' and 'obs' items in the output dictionary were renamed to 'ydata_mdl' and 'ydata_obs', respectively.
- For output of differences: an 'xdiff' item was added to the output dictionary; and the 'diff' item in the output dictionary was renamed to 'ydiff'.
