### 2022-05-19 #4326 @slayoo
Here is some more feedback on the installation instructions:
- worth mentioning that the `conda` steps are optional, and one can use the package without `conda`
- the `conda install --file requirements.txt` seem unneeded now that `setup.cfg` contains the dependency info
- the "To run the tests..." subsection does not contain any information on how to run the tests - just test data retrieval
- there are two broken links in the "Examples" section (under "configuration" and "installation")
