### 2022-05-21 64a80ff
update 'statement of need' section of paper.md to clarify software package distinctions
