### 2022-05-21 #4326 @dkauf42
> @dkauf42, regarding the substantial scholarly effort section of the review.
>
> The ESMValTool is rightly referenced in the statement of need in the paper. And its overarching objective appears to be the same as gdess, which means the benefit of gdess (over ESMValTool) lies in the details (but tell me if I'm wrong on this).
>
> The distinction between the two software packages is stated in the gdess paper (Statement of need section) as: '—but does not provide for tailored processing of varied CO2 data sources'. From this statement I cannot tell the benefit of gdess over ESMValTool.
>
> Although I appreciate the need for concision in the JOSS paper, I recommend that the advantage of gdess over ESMValTool be expanded in the Statement of need, so that it is clearer.

_@simonom, to help clarify the distinction between these software packages, we've revised and added text to the 2nd paragraph of the 'statement of need' section (commit [64a80ff](https://github.com/E3SM-Project/gdess/commit/64a80ff697d10d3e7bcc9029b3ae53e446441246)). The revised text reads:_

"For example, the ESM Evaluation Tool (ESMValTool; @Eyring2016c; @Eyring2020b) has been used to generate specific figures from the literature, and we adopted the term *recipe* from its use by ESMValTool. Although ESMValTool includes a comparison of column-averaged CO<sub>2</sub> values as performed by Gier et al. (2020), `gdess` was created to provide specific CO<sub>2</sub> diagnostic methods and graphs that are not already provided as recipes in ESMValTool. `gdess` uses Observation Package (Obspack; @globalview20200911; @Masarie2014b) data, which include atmospheric greenhouse gas observations from a variety of sampling platforms and data providers following the World Data Centre for Greenhouse Gases (WDCGG) protocol, so are widely used for stimulating and supporting carbon cycle modeling studies.  These data have not been set up for use within ESMValTool, and as such would require additional development/configuration to work with ESMValTool."
