### 2022-05-21 3e8db1d
fix description of using `git lfs` how to run the built-in tests in readme
