### 2022-05-22 #4326 @dkauf42
> Here is some more feedback on the installation instructions:
>
> * **(a)** worth mentioning that the `conda` steps are optional, and one can use the package without `conda`
> * **(b)** the `conda install --file requirements.txt` seem unneeded now that `setup.cfg` contains the dependency info
> * **(c)** the "To run the tests..." subsection does not contain any information on how to run the tests - just test data retrieval
> * **(d)** there are two broken links in the "Examples" section (under "configuration" and "installation")

@slayoo, we appreciate the detailed feedback, and here respond to each of these points:

- **(a)** We have repeatedly run into errors installing `cartopy`, one of `gdess`'s dependencies, with just 'pip' and 'setup.cfg'.  Installing dependencies via `conda` has enabled us to avoid this issue, and that is why we have kept the `conda` instructions in the readme.
- **(b)** We understand that this seems a bit redundant. Although we would like to reduce duplication to a minimum, we have kept this step in the README for the same reason as in point (a) above.
- **(c)** This is a good point. To avoid further confusion, we have created a separate "Running the code tests" section in the readme, and clarified the language in the installation section (see commit [3e8db1](https://github.com/E3SM-Project/gdess/pull/9/commits/3e8db1dfb9f8de2c2da2b032d39b06ea040a2602)).
- **(d)** These links have been corrected with commit [8f7ec0](https://github.com/E3SM-Project/gdess/pull/9/commits/8f7ec0d2756d3dc81c41aae5929f6f702c858062).
