### 2022-05-30 #4326 @simonom
@dkauf42, I realised that the configuration instructions inside the .bat file (for Windows systems) were missing some detail, so have updated in the attached, and recommend inclusion in the software.  PLEASE NOTE, I couldn't attach a .bat file here, so have changed to a .txt file, and the extension therefore needs reverting back to .bat
[set_path_vars.txt](https://github.com/openjournals/joss-reviews/files/8797542/set_path_vars.txt)
