### 2022-06-11 #4326 @slayoo
First of all, apologies for the delay.

Some final remarks:
- please replace raster graphics with vector graphics in the paper, this will also attest to the ability of `gdess` to produce publication-quality output (hence, as a placeholder, leaving out "Quality of writing" checkbox unchecked for now)
- while the paper mentions comprehensive docstrings, there seem to be no instructions provided for how to best view the docstring content - perhaps worth pointing out the users to tools such as `pdoc`, `sphinx` or alike and checking how the documentation is rendered there, optimally it would be great to have it published on every merge to the main branch somewhere on the github web space (currently, the github website for `gdess` is bugus: https://e3sm-project.github.io/gdess/). Leaving out "Functionality documentation" checkbox unchecked for now.

On the maintenance side:
- CI is currently done only for `ubuntu-latest`, highly suggest maintaining it also for other targeted platforms (i.e., `macos-latest` and `windows-latest`), especially as instructions are given in the README file for macOS and Windows
- highly suggest adding a Github Actions workflow which runs the demo notebooks ensuring everything still works while adding new features to the package (essentially, something like: `for i in notebooks/demo/*.ipynb; do python -m nbconvert --to notebook --execute $i; done;` ... which gives `jupyter_client.kernelspec.NoSuchKernel: No such kernel named test_gdess_for_joss` on my machine :( - some hardcoded kernel label to be removed, see e.g. here: https://github.com/E3SM-Project/gdess/blob/e3f65b0b92be21b2bf9de923c2180dd8d8696192/notebooks/demo/meridional_recipe.ipynb?short_path=c568465#L519)

Last but not least, when attempting to execute the tests with Python 3.8.7, I got the following warnings, which are worth addressing (likely by using raw strings?):
```
gdess/operations/convert.py:10
  /home/slayoo/devel/gdess/gdess/operations/convert.py:10: DeprecationWarning: invalid escape sequence \,
    """Convert CO2 from units of mol/mol to ppm

gdess/operations/convert.py:54
  /home/slayoo/devel/gdess/gdess/operations/convert.py:54: DeprecationWarning: invalid escape sequence \,
    """Convert CO2 from units of kg/kg to ppm

tests/test_curvefitting.py:34
  /home/slayoo/devel/gdess/tests/test_curvefitting.py:34: DeprecationWarning: invalid escape sequence \s
    df_filter_output = pd.read_csv(output_file_name, sep='\s+')

tests/test_curvefitting.py:41
  /home/slayoo/devel/gdess/tests/test_curvefitting.py:41: DeprecationWarning: invalid escape sequence \s
    df_expected = pd.read_csv(expected_results_path, sep='\s+')

gdess/data_source/models/e3sm/calculation.py:18
  /home/slayoo/devel/gdess/gdess/data_source/models/e3sm/calculation.py:18: DeprecationWarning: invalid escape sequence \D
    """Change in mass of $CO_2$ from timestep to timestep

gdess/data_source/models/e3sm/calculation.py:36
  /home/slayoo/devel/gdess/gdess/data_source/models/e3sm/calculation.py:36: DeprecationWarning: invalid escape sequence \s
    """Global means with weighting by area

gdess/data_source/models/e3sm/calculation.py:60
  /home/slayoo/devel/gdess/gdess/data_source/models/e3sm/calculation.py:60: DeprecationWarning: invalid escape sequence \i
    """Time-integrated flux of $CO_2$ from surface emissions and aircraft

tests/test_formatters.py:29
  /home/slayoo/devel/gdess/tests/test_formatters.py:29: DeprecationWarning: invalid escape sequence \$
    assert tex_escape('\$ \_ >') == \
```

HTH
