### 2022-07-06 #4326 @simonom
Hiya,

In addition to Slayoo's points, I have two (possibly final, but definitely nearly final) points:

i) Inside the installation instructions on the README, I suggest a link to the conda website (to help users without conda get setup)

ii) Inside the Motivation section of the README, I suggest an edit to make the GitHub page more welcoming to users from outside the E3SM community (since gdess is applicable to other data sources)
