### 2022-07-16 8917ab3
update figure output with svg version
