### 2022-07-16 f2e6e00
replace paper figures with pdf version
