### 2022-07-16 c591f3c
update docstrings
