+++ gdess/operations/Confrontation.py
@@ -1,5 +1,15 @@
+import argparse
+import csv, sys, logging
+from typing import Union, Iterable
+from datetime import datetime
+
+import numpy as np
+import pandas as pd
+import xarray as xr
+from dask.diagnostics import ProgressBar
+from sklearn.metrics import mean_squared_error
+
 import gdess.graphics
-from ccgcrv.ccg_filter import ccgFilter
 from gdess import set_verbose
 from gdess.data_source.models.cmip.cmip_collection import Collection as cmipCollection
 from gdess.graphics.single_source_plots import plot_filter_components
@@ -8,15 +18,8 @@ from gdess.operations.geographic import get_closest_mdl_cell_dict
 from gdess.operations.utils import assert_expected_dimensions
 from gdess.formatters import append_before_extension
 from gdess.data_source.observations import gvplus_surface as obspack_surface_collection_module
+from ccgcrv.ccg_filter import ccgFilter
 from ccgcrv.ccg_dates import decimalDateFromDatetime
-from sklearn.metrics import mean_squared_error
-from datetime import datetime
-import numpy as np
-import pandas as pd
-import xarray as xr
-from dask.diagnostics import ProgressBar
-from typing import Union
-import csv, sys, logging
 
 _logger = logging.getLogger(__name__)
 
@@ -24,8 +27,8 @@ _logger = logging.getLogger(__name__)
 class Confrontation:
     def __init__(self,
                  compare_against_model: bool,
-                 ds_mdl,
-                 opts,
+                 ds_mdl: xr.Dataset,
+                 opts: argparse.Namespace,
                  stations_to_analyze: list,
                  verbose: Union[bool, str] = False):
         """Instantiate a Confrontation object.
@@ -46,8 +49,8 @@ class Confrontation:
 
         set_verbose(_logger, verbose)
 
-    def looper(self, how):
-        """
+    def looper(self, how: str) -> tuple:
+        """Process and format observation data for multiple station locations
 
         Parameters
         ----------
@@ -61,7 +64,9 @@ class Confrontation:
         Returns
         -------
         tuple
-            A bunch of things
+            data_dict, concatenated_dfs, df_station_metadata, \
+               xdata_gv, xdata_mdl, ydata_gv, ydata_mdl, \
+               rmse_y_true, rmse_y_pred
         """
         valid = {'seasonal', 'trend'}
         if how not in valid:
@@ -249,7 +254,10 @@ class Confrontation:
                xdata_gv, xdata_mdl, ydata_gv, ydata_mdl, \
                rmse_y_true, rmse_y_pred
 
-    def concatenate_stations_and_months(self, data_dict, processed_station_metadata) -> (dict, pd.DataFrame):
+    def concatenate_stations_and_months(self,
+                                        data_dict: dict,
+                                        processed_station_metadata: dict
+                                        ) -> (dict, pd.DataFrame):
         """
 
         Parameters
@@ -287,7 +295,10 @@ class Confrontation:
         return df_concatenated, df_station_metadata
 
 
-def make_comparable(ref: xr.Dataset, com: xr.Dataset, **keywords) -> (xr.Dataset, xr.DataArray):
+def make_comparable(ref: xr.Dataset,
+                    com: xr.Dataset,
+                    **keywords
+                    ) -> (xr.Dataset, xr.DataArray):
     """Make two datasets comparable.
 
     Ensures time formats are compatible.
@@ -321,7 +332,6 @@ def make_comparable(ref: xr.Dataset, com: xr.Dataset, **keywords) -> (xr.Dataset
         the modified reference variable object
     com : xarray.Dataarray
         the modified comparison variable object
-
     """
 
     # Process keywords
[…]
