### 2022-07-16 e4e6fcb
update __all__ and module docstrings
