### 2022-07-16 8734629
add script for creating doc htmls
