### 2022-07-17 58c93ce
use latex math option in pdoc template
