### 2022-07-17 f74bdcd
improve docstrings, type hints, and update html docs
