### 2022-07-17 7910f52
add top level package docstring
