### 2022-07-17 9d2c1dd
docstring tweak
