+++ tests/conftest.py
@@ -1,7 +1,9 @@
 import os
+from pathlib import Path
 import pytest
 
 
 @pytest.fixture
 def rootdir():
-    return os.path.dirname(os.path.abspath(__file__))
+    return Path(__file__).parent.resolve()
+        # os.path.dirname(os.path.abspath(__file__))
