+++ tests/test_arg_utils.py
@@ -1,5 +1,5 @@
 import pytest
-import os
+from pathlib import Path
 
 from gdess.formatters.args import options_to_args, is_some_none, nullable_int, nullable_str, valid_year_string, \
     valid_existing_path, valid_writable_path
@@ -75,11 +75,11 @@ def test_nullable_integer_bad_type():
         nullable_int([])
 
 
-def test_nonexistent_path_does_not_exist(rootdir):
+def test_nonexistent_path_does_not_exist(rootdir: Path):
     with pytest.raises(Exception):
-        valid_existing_path(os.path.join(rootdir, 'nonexistent'))
+        valid_existing_path(rootdir / 'nonexistent')
 
 
-def test_valid_path_is_writable(rootdir):
-    p = os.path.join(rootdir, 'footest')
+def test_valid_path_is_writable(rootdir: Path):
+    p = rootdir / 'footest'
     assert valid_writable_path(p) == p
