+++ tests/test_obspack_surface_collection_recipes.py
@@ -1,10 +1,17 @@
+import os
+import re
 import pytest
 from pathlib import Path
 
+import numpy as np
 import xarray as xr
 
 from gdess import load_stations_dict
+from gdess.data_source.observations.load import load_data_with_regex
+from gdess.data_source.observations.subset import binTimeLat, binLonLat, by_datetime
 from gdess.data_source.observations.gvplus_surface import Collection
+from gdess.data_source.observations.gvplus_name_utils import \
+    get_dict_of_all_station_filenames, get_dict_of_station_codes_and_names
 
 
 @pytest.fixture
@@ -16,10 +23,39 @@ def newEmptySurfaceStation():
 def globalview_test_data_path(rootdir: Path):
     return rootdir / 'test_data' / 'globalview'
 
+@pytest.fixture
+def globalview_datasetdict(globalview_test_data_path):
+    p = re.compile(r'co2_([a-zA-Z0-9]*)_surface.*\.nc$')
+    return load_data_with_regex(globalview_test_data_path, compiled_regex_pattern=p)
+
 def test_station_MLO_is_present(newEmptySurfaceStation):
     station_dict = load_stations_dict()
     assert 'mlo' in station_dict
 
+def test_station_MLO_in_all_station_filenames(globalview_test_data_path):
+    a = get_dict_of_all_station_filenames(str(globalview_test_data_path) + os.sep)
+    assert 'mlo' in a.keys()
+
+def test_station_MLO_in_all_station_datasets(globalview_test_data_path):
+    a = get_dict_of_station_codes_and_names(str(globalview_test_data_path) + os.sep)
+    assert 'mlo' in a.keys()
+
+def test_station_MLO_in_loaded_datasets(globalview_datasetdict):
+    assert 'mlo' in globalview_datasetdict.keys()
+
+def test_return_type_from_timelatbinned_dataset(globalview_datasetdict):
+    binned = binTimeLat(globalview_datasetdict['mlo'])
+    assert isinstance(binned, tuple)
+
+def test_return_type_from_lonlatbinned_dataset(globalview_datasetdict):
+    binned = binLonLat(globalview_datasetdict['mlo'])
+    assert isinstance(binned, tuple)
+
+def test_return_type_from_datetime_binning(globalview_datasetdict):
+    binned = by_datetime(globalview_datasetdict['mlo'],
+                         start=np.datetime64("2000-01-01"),
+                         end=np.datetime64("2000-03-01"))
+    assert isinstance(binned, xr.Dataset)
 
 def test_simplest_preprocessed_type(globalview_test_data_path, newEmptySurfaceStation):
     newEmptySurfaceStation.preprocess(datadir=globalview_test_data_path, station_name='mlo')
