+++ tests/conftest.py
@@ -1,4 +1,3 @@
-import os
 from pathlib import PurePath
 import pytest
 
@@ -6,4 +5,3 @@ import pytest
 @pytest.fixture
 def rootdir():
     return PurePath(__file__).parent
-        # os.path.dirname(os.path.abspath(__file__))
