### 2022-07-18 3624b44
use PurePath exists() for validation
