### 2022-07-18 0e29095
clean print statements
