### 2022-07-18 cb67565
convert posix path to purepath before validating
