### 2022-07-18 6a58411
convert purepath to path
