### 2022-07-18 6c937ae
rework path validation error checking
