### 2022-07-18 357d6f8
use pytest fixture for output dir
