### 2022-07-19 bd542e7
consolidate test paths
