+++ tests/test_arg_utils.py
@@ -1,5 +1,4 @@
 import pytest
-from pathlib import Path
 
 from gdess.formatters.args import options_to_args, is_some_none, nullable_int, nullable_str, valid_year_string, \
     valid_existing_path, valid_writable_path
