### 2022-07-19 30ae044
consolidate test paths
