### 2022-07-19 2ffd826
add testing for platform on arg parsing
