### 2022-07-19 df3971f
use posix False flag on windows for shlex split, which otherwise swallows backslashes
