### 2022-07-22 f8614a0
fix notebook paths
