### 2022-07-23 cf93d35
use nbconvert on each demo notebook template
