+++ .github/workflows/lint_and_test.yml
@@ -123,3 +123,14 @@ jobs:
         with:
           notebook: "notebooks/demo/meridional_recipe.ipynb"
 
+      - name: Update analysis notebook
+        run: |
+          python -m pip install jupyter nbconvert nbformat
+          for i in notebooks/templates/*recipe.ipynb; do python -m nbconvert --to notebook --ExecutePreprocessor.kernel_name=python3 --execute $i
+      - name: commit updated seasonal notebook
+        uses: EndBug/add-and-commit@v7
+        with:
+          author_name: Plot update bot
+          message: "Update Notebook"
+          add: "seasonal_cycle_recipe.ipynb"
+
