### 2022-07-23 970a7c1
try fixing the nbconvert command and use latest add-and-commit version
