+++ .github/workflows/lint_and_test.yml
@@ -110,14 +110,14 @@ jobs:
           fail_ci_if_error: false
           verbose: true # optional (default = false)
 
-      - name: Update analysis notebook
+      - name: Update demo notebooks
         run: |
           python -m pip install jupyter nbconvert nbformat
-          for i in notebooks/templates/*recipe.ipynb; do python -m nbconvert --to notebook --ExecutePreprocessor.kernel_name=python3 --execute $i
-      - name: commit updated seasonal notebook
-        uses: EndBug/add-and-commit@v7
+          for i in notebooks/templates/*recipe.ipynb; do python -m nbconvert --to notebook --ExecutePreprocessor.kernel_name=python3 --execute $i; done;
+      - name: Commit updated notebooks
+        uses: EndBug/add-and-commit@v9
         with:
           author_name: Plot update bot
-          message: "Update Notebook"
-          add: "seasonal_cycle_recipe.ipynb"
+          message: "Update Notebooks"
+          add: "['meridional_recipe.ipynb', 'seasonal_cycle_recipe.ipynb', 'trends_recipe.ipynb']"
 
