### 2022-07-23 acad0b5
use login shell in actions workflow
