### 2022-07-23 2ae75d4
use "_template" to distinguish pre-actions notebook file and handle filename in bash command
