+++ .github/workflows/lint_and_test.yml
@@ -114,7 +114,7 @@ jobs:
         shell: bash -l {0}
         run: |
           python -m pip install jupyter nbconvert nbformat
-          for i in notebooks/templates/*recipe.ipynb; do python -m nbconvert --to notebook --ExecutePreprocessor.kernel_name=python3 --execute $i; done;
+          for i in notebooks/templates/*recipe_template.ipynb; do filename=${i##*/}; python -m nbconvert --to notebook --ExecutePreprocessor.kernel_name=python3 --execute $i --output ${filename%_template.ipynb}.ipynb; done;
       - name: Commit updated notebooks
         uses: EndBug/add-and-commit@v9
         with:
