### 2022-07-23 0e8b8fd
fix paths to converted notebooks
