### 2022-07-23 22d1182
checkout the branch the PR is coming from during a run of GitHub Actions, so that we're not in a detached head state for later commits
