+++ .github/workflows/lint_and_test.yml
@@ -35,6 +35,8 @@ jobs:
       - name: Checkout source
         uses: actions/checkout@v2
         with:
+          repository: ${{ github.event.pull_request.head.repo.full_name }}
+          ref: ${{ github.event.pull_request.head.ref }}
           lfs: true
         #      - name: Set up Python ${{ matrix.python-version }}
         #        uses: actions/setup-python@v2
@@ -115,6 +117,7 @@ jobs:
         run: |
           python -m pip install jupyter nbconvert nbformat
           for i in notebooks/templates/*recipe_template.ipynb; do filename=${i##*/}; python -m nbconvert --to notebook --ExecutePreprocessor.kernel_name=python3 --execute $i --output ${filename%_template.ipynb}.ipynb; done;
+
       - name: Commit updated notebooks
         uses: EndBug/add-and-commit@v9
         with:
