### 2022-07-23 26cabf0
add os and python conditional to limit notebook running and committing
