+++ .github/workflows/lint_and_test.yml
@@ -22,7 +22,7 @@ jobs:
     name:  Test1 (${{ matrix.python-version }}, ${{ matrix.os }})
     runs-on: ${{ matrix.os }}
     strategy:
-      fail-fast: false
+      fail-fast: true
       matrix:
         os: [ubuntu-latest, macOS-latest, windows-latest]
         python-version: [3.8, 3.9]
@@ -113,12 +113,14 @@ jobs:
           verbose: true # optional (default = false)
 
       - name: Update demo notebooks
+        if: (matrix.os == 'ubuntu-latest') && (matrix.python-version == '3.9')
         shell: bash -l {0}
         run: |
           python -m pip install jupyter nbconvert nbformat
           for i in notebooks/templates/*recipe_template.ipynb; do filename=${i##*/}; python -m nbconvert --to notebook --ExecutePreprocessor.kernel_name=python3 --execute $i --output ${filename%_template.ipynb}.ipynb; done;
 
       - name: Commit updated notebooks
+        if: (matrix.os == 'ubuntu-latest') && (matrix.python-version == '3.9')
         uses: EndBug/add-and-commit@v9
         with:
           author_name: Plot update bot
