### 2022-07-24 990bbc4
use output-dir argument for nbconvert
