+++ .github/workflows/lint_and_test.yml
@@ -117,8 +117,8 @@ jobs:
         shell: bash -l {0}
         run: |
           python -m pip install jupyter nbconvert nbformat
-          for i in notebooks/templates/*recipe_template.ipynb; do filename=${i##*/}; python -m nbconvert --to notebook --ExecutePreprocessor.kernel_name=python3 --execute $i --output notebooks/demo/${filename%_template.ipynb}.ipynb; done;
-
+          for i in notebooks/templates/*recipe_template.ipynb; do filename=${i##*/}; python -m nbconvert --to notebook --ExecutePreprocessor.kernel_name=python3 --execute $i --output-dir notebooks/demo --output ${filename%_template.ipynb}.ipynb; done;
+--output-dir notebooks/demo
       - name: Commit updated notebooks
         if: (matrix.os == 'ubuntu-latest') && (matrix.python-version == '3.9')
         uses: EndBug/add-and-commit@v9
