+++ README.md
@@ -1,8 +1,8 @@
 # Greenhouse gas Diagnostics for Earth System Simulations
 
 <!-- Information badges -->
-[![test status](https://github.com/dkauf42/gdess/actions/workflows/lint_and_test.yml/badge.svg)](https://github.com/dkauf42/gdess/actions/workflows/lint_and_test.yml)
-[![codecov](https://codecov.io/gh/dkauf42/gdess/branch/main/graph/badge.svg?token=UFKDMQ9KCM)](https://codecov.io/gh/dkauf42/gdess)
+[![test status](https://github.com/E3SM-Project/gdess/actions/workflows/lint_and_test.yml/badge.svg)](https://github.com/E3SM-Project/gdess/actions/workflows/lint_and_test.yml)
+[![codecov](https://codecov.io/gh/E3SM-Project/gdess/branch/main/graph/badge.svg?token=UFKDMQ9KCM)](https://codecov.io/gh/E3SM-Project/gdess)
 [![Python 3.8|3.9](https://img.shields.io/badge/python-3.8|3.9-blue.svg)](https://www.python.org/downloads/release/python-380/)
 [![status](https://joss.theoj.org/papers/6cc7ba8d3c09f66974550161f740e176/status.svg)](https://joss.theoj.org/papers/6cc7ba8d3c09f66974550161f740e176)
 
