### 2022-07-24 f175e64
add link to conda docs in the readme installation section
