### 2022-07-24 0956f16
revise 'motivation' section to include a broader description of reasons for comparing ESMs against observations
