### 2022-07-24 #4326 @dkauf42
Dear @slayoo, once again thank you for your comments and recommendations. Our responses are below, and you will find the associated changes in [Pull Request 13](https://github.com/E3SM-Project/gdess/pull/13):
-	We have replaced the `gdess` output figures (3,4, and 5) with vector (.svg) versions.
-	We have completely revised the github.io page (https://e3sm-project.github.io/gdess/) to show documentation that has been auto-generated using `pdoc`. Now, GitHub re-publishes that page using the `docs/` directory in the repo. And the html pages in `docs` can be re-generated by running the newly included `build.sh` script in `doc/`.
-	The CI workflow now runs—and is passing—for `macos-latest` and windows-latest`, in addition to `ubuntu-latest`.
-	Re-generating the notebooks in the CI workflow is a great idea. To implement this, we had to split the meridional and seasonal cycle notebooks into two parts: one that can be executed using `nbconvert` in the GitHub actions workflow using the test data contained in the repo, and another part that must remain static because it requires manual downloading of many observation data files. The auto-executed notebooks have been placed in `notebooks/demo/` and you will notice that the two new static notebooks are there as well, with the suffix “_requiring_download.ipynb”.  The workflow uses `nbconvert` to execute the notebooks from template files in `notebooks/templates` and auto-commits the new notebooks.
- Strings that were causing the "invalid escape sequence" warnings have been replaced with raw strings, and the warnings are not appearing.

Dear @simonom, we thank you for your points regarding the README, which we believe will help future readers/developers. To address your comments, the installation section and motivation section of the README have been revised, in commits [0956f16](https://github.com/E3SM-Project/gdess/pull/13/commits/0956f163d2f311ebfe78c60c8dff0290c41478e8) and [f175e64](https://github.com/E3SM-Project/gdess/pull/13/commits/f175e64a43910d5f869a6470fb04c7a12ad00fa0). A link to `conda` is now included in the installation section, and the motivation section now includes a broader description of reasons for comparing Earth System Models against observations.
