### 2022-07-26 #4326 @simonom
Thanks @dkauf42, all my checkboxes are now ticked.  @dhhagan, I recommend publication in JOSS.
