### 2022-07-26 #4326 @slayoo
@dhhagan, same here - I recommend publication in JOSS.
