### 2022-08-10 #4326 @dkauf42
@dhhagan, sure thing! Though just to double-check, do you mean archiving `gdess` with something like [zenodo.org](https://zenodo.org/)?
