+++ setup.cfg
@@ -5,7 +5,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 author = Daniel E Kaufman
 author_email = daniel.kaufman@pnnl.gov
-version = 1.0.beta1
+version = 1.0
 classifiers =
   Intended Audience :: Science/Research
   Programming Language :: Python :: 3.8
