### 2022-08-26 31d51d6
fix isbn issue with reference in .bib
