+++ paper/paper.bib
@@ -305,13 +305,7 @@
 	number       = 10,
 	pages        = {5155--5190},
 	doi          = {10.1002/2014JD022591},
-	isbn         = {2169-9356},
-	issn         = {2169897X},
 	url          = {http://doi.wiley.com/10.1002/2014JD022591},
-	archiveprefix = {arXiv},
-	arxivid      = {arXiv:1402.6991v1},
-	eprint       = {arXiv:1402.6991v1},
-	pmid         = 23641116,
 }
 @article{Thoning1989jgr,
 	title        = {{Atmospheric carbon dioxide at Mauna Loa Observatory. 2. Analysis of the NOAA GMCC data, 1974-1985}},
