### 2022-08-26 #4326 @dkauf42
@dhhagan, I saw this error a few minutes ago, and I think it's because there was an ISBN listed for Sweeney et al. (2015) in the ".bib" file that shouldn't have been there. I just removed it and pushed the change to the `main` branch.
Can we try this process again?
