### 2022-08-29 #4326 @slayoo
Hello, @dhhagan, @dkauf42, @arfon!
We have a problem here with the LaTeX tilde visible in the paper title on the JOSS website.

![image](https://user-images.githubusercontent.com/89685/187141606-861b3049-8223-4279-9696-52fe8a2ec5db.png)

Also in "citation string":
![image](https://user-images.githubusercontent.com/89685/187141749-fe1fc610-a1bb-43f7-9a8e-3456061b73ec.png)

Same in BiBTeX metadata.

And on Twitter:
![image](https://user-images.githubusercontent.com/89685/187142004-645c405c-1f16-4e2d-9aa7-c34f7e00fe3b.png)
