### 2022-08-29 #4326 @dkauf42
I'm not sure what to suggest for this. The tildes were used to render the subscript appropriately in the `paper.md` markdown file.
Perhaps there needs to be additional text parsing in the automated process for generating webpage/BiBTeX/twitter text?
